chore: Move tests to a dedicated tests folder chore: Replace array of encumbrance steps with a more descriptive object
