+++ src/e2e/index.ts
@@ -10,27 +10,27 @@ import characterTests, {
 import dataModelCharacterACTests, {
   key as dataModelCharacterACKey,
   options as dataModelCharacterACOptions
-} from '../module/actor/data-model-classes/data-model-character-ac.test';
+} from '../module/actor/data-model-classes/__tests__/data-model-character-ac.test';
 
 import dataModelCharacterScoresTests, {
   key as dataModelCharacterScoresKey,
   options as dataModelCharacterScoresOptions
-} from '../module/actor/data-model-classes/data-model-character-scores.test';
+} from '../module/actor/data-model-classes/__tests__/data-model-character-scores.test';
 
 import dataModelCharacterSpellsTests, {
   key as dataModelCharacterSpellsKey,
   options as dataModelCharacterSpellsOptions
-} from '../module/actor/data-model-classes/data-model-character-spells.test';
+} from '../module/actor/data-model-classes/__tests__/data-model-character-spells.test';
 
 import dataModelCharacterEncumbranceTests, {
   key as dataModelCharacterEncumbranceKey,
   options as dataModelCharacterEncumbranceOptions
-} from '../module/actor/data-model-classes/data-model-character-encumbrance.test';
+} from '../module/actor/data-model-classes/__tests__/data-model-character-encumbrance.test';
 
 import dataModelCharacterMoveTests, {
   key as dataModelCharacterMoveKey,
   options as dataModelCharacterMoveOptions
-} from '../module/actor/data-model-classes/data-model-character-move.test';
+} from '../module/actor/data-model-classes/__tests__/data-model-character-move.test';
 
 import dataModelCharacterTests, {
   key as dataModelCharacterKey,
