+++ src/module/actor/data-model-classes/__tests__/data-model-character-scores.test.ts
@@ -1,5 +1,5 @@
-import { QuenchMethods } from "../../../e2e";
-import OseDataModelCharacterScores from "./data-model-character-scores";
+import { QuenchMethods } from "../../../../e2e";
+import OseDataModelCharacterScores from "../data-model-character-scores";
 
 export const key = 'ose.datamodel.character.scores';
 export const options = { displayName: 'Character Data Model: Ability Scores'}
